+++ packages/backend/convex/github.ts
@@ -10,7 +10,7 @@ export {
 
 export { createSessionPr, createDraftSessionPr } from "./_github/prFlow";
 
-export { getPrDiff } from "./_github/prDiff";
+export { getPrDiff, getPrFileContents } from "./_github/prDiff";
 
 export { listPullRequests, getPullRequestHeader } from "./_github/pullRequests";